Multiprocessing pool
Modified main function using a multiprocessing pool.

Depending on the size of the input files and the number of available CPU cores, it may be faster to use a multiprocessing pool instead of a concurrent.futures thread pool. This would allow the processing to be parallelized across multiple processes instead of just multiple threads.